requirements.txt
Flask
beautifulsoup4
requests
Werkzeug
PyQt5
GitPython
python-dotenv